backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
aiokafka          # Async Kafka for FastAPI
confluent-kafka   # High-performance Kafka for the Consumer
clickhouse-connect
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
asyncpg
gunicorn
pytest
pytest-asyncio
httpx
stripe
